# src/main/resources/templates/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger lokapi - OpenAPI 3.0
  description: ''
  termsOfService: http://swagger.io/terms/
  contact:
    email: tawfik.elyousfi@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://test.swagger.io/api/v3
...
paths:
...
  /users:
    get:
      tags:
        - user
      summary: Get all Users
      operationId: getAllUsers
      responses:
        default:
          description: List of Users received with success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '404':
          description: Users not found
    post:
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
        - user
      summary: update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      requestBody:
        description: update user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      requestBody:
        description: Deleted user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      tags:
        - user
      summary: Get user by user Id
      description: ''
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: 'The id that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUserById
      parameters:
        - name: userId
          in: path
          description: ID that need to be updated
          required: true
          schema:
            type: integer
            format: int64
      #requestBody:
      #  description: Update an existent user in the store
      #  content:
      #    application/json:
      #      schema:
      #        $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUserById
      parameters:
        - name: userId
          in: path
          description: The ID that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
  /users/{username}:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
  /users/{email}:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByEmail
      parameters:
        - name: email
          in: path
          description: 'The email that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid email supplied
        '404':
          description: User not found
  /users/{emailOrUsername}:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByIdentifiant
      parameters:
        - name: emailOrUsername
          in: path
          description: 'The email or username that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid email supplied
        '404':
          description: User not found
  /users/{existingEmailOrUsername}:
    get:
      tags:
        - user
      summary: Check if user is registered
      description: ''
      operationId: isUserByIdentifiantRegistered
      parameters:
        - name: existingEmailOrUsername
          in: path
          description: 'The email or username that needs to be fetched. check if user is existing. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: boolean
  /users/{ownerId}:
    get:
      tags:
        - user
      summary: Get tenant from an owner
      description: ''
      operationId: getTenantByOwner
      parameters:
        - name: ownerId
          in: path
          description: 'Ge all tenants of the owner logged '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Invalid email supplied
        '404':
          description: User not found
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: 'theUser'
        email:
          type: string
          example: 'you@email.com'
        password:
          type: string
          example: '12345'
        roles:
          type: array
          items:
            type: string
            enum:
              - OWNER
              - TENANT
          example:
            - TENANT
            - OWNER
          description: tenant or owner status in application

        active:
            type: boolean
            example: 1
        creationDate:
          type: string
          format: date
          description: User creation date
        updateDate:
          type: string
          format: date
          description: User update date
        deleteDate:
          type: string
          format: date
          description: User detele date
      required:
        - id
        - username
        - email
        - password
        - roles
...